Fixed: Changed Main page's Layout & Route Domain
