+++ frontend/src/pages/mainPage/mainPage.module.css
@@ -1,0 +1,3 @@
+.container {
+  user-select: none;
+}
