+++ frontend/src/pages/mainPage/selectArea/allArea/allArea.jsx
