+++ frontend/src/pages/mainPage/selectArea/allArea/allArea.module.css
