+++ frontend/src/pages/mainPage/selectArea/selectArea.jsx
