+++ frontend/src/pages/mainPage/selectArea/selectArea.module.css
