--- frontend/src/pages/selectContainer/selectContainer.module.css
@@ -1,3 +1,0 @@
-.container {
-  user-select: none;
-}
